Tag management in the API
After a previous approach with relational tables (Image, Tag,
BoundingBox) I decided that it was too complex and also required too
many SQL statements for each updated tag.
So I decided to use a JSON field in images to store the bounding boxes,
and a second table that stores Tags, with a JSON field that references
all image_ids that are using the tag.