Bugfix for bounding boxes dragging, and minor bugfix for image select
When the user is dragging a box, the POST for tags update is not
executed.
But if the user started dragging just after the request was already
initiated, the response to getTags would cause an update of the bounding
boxes that caused a Konva crash because the bboxes were removed while
being dragged.